requirements.txt
streamlit==1.29.0
boto3==1.34.0
reportlab==4.0.7
python-pptx==0.6.23
pandas==2.1.4
plotly==5.17.0